backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.1

# Database
sqlmodel==0.0.22
asyncpg==0.31.0

# Validation and Settings
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.1.0

# Authentication and Security
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.6

# AI/ML - Phase III
openai==1.57.4
